+++ app/Http/Livewire/Reports/SpecialCurProgram/SPA/SPAReport.php
@@ -3,6 +3,7 @@
 namespace App\Http\Livewire\Reports\SpecialCurProgram\SPA;
 
 use App\Models\SPAModel;
+use Illuminate\Support\Facades\Hash;
 use Livewire\Component;
 use Livewire\WithPagination;
 
@@ -11,13 +12,24 @@ class SPAReport extends Component
     use WithPagination;
 
     public $search = '';
+    protected $deleteIdentifier = 'DELETESPA';
 
-    public function deleteSPA($id){
+    protected $listeners = ['deleteRecord'];
+
+    public function deleteRecord($id, $componentIdentifier){
+        if(!Hash::check($this->deleteIdentifier, $componentIdentifier)){
+            return;
+        }
          // Find the Academic Track by ID
         $SPA = SPAModel::findOrFail($id);
         
         // Delete the Academic Track
         $SPA->delete();
+
+         $this->emit('showNotifications', [
+        'type' => 'success',
+        'message' => 'Record Deleted',
+    ]);
     }
 
     public function navigateTo($link, $id)
@@ -30,6 +42,12 @@ class SPAReport extends Component
     {
         $this->resetPage();
     }
+
+    protected function hashedDeleteIdentifier()
+    {
+        return Hash::make($this->deleteIdentifier);
+    }
+
     
     public function render()
     {
@@ -40,7 +58,8 @@ class SPAReport extends Component
                     ->orWhere('name', 'like', '%' . $this->search . '%');
             })
             ->orWhere('grade_lvl', 'like', '%' . $this->search . '%')
-            ->paginate(5)
+            ->paginate(5),
+            'deleteIdentifier' => $this->hashedDeleteIdentifier(),
         ]);
     }
 }
